finalized v1.0

+++ routes/gifts.js
@@ -1,21 +1,21 @@
 var express = require('express');
 var router = express.Router();
 var Gift = require('../models/gift.model');
+var People = require('../models/people.model');
 
-// Add a user
+// Add a gift
 router.post('/people/:id/gifts', function (req, res) {
-    var gift = new Gift(req.body);
-    gift.save(function (err, gift) {
-        console.log(gift);
-        if (err) return res.status(500).json({ msg: 'Error adding gift.', data: err });
-        return res.status(200).json({ data: gift });
+    //Add gift and the gift id to the person
+    People.findById(req.params.id, (err, person) => {
+        if (err || person==null) return res.status(500).json({ msg: 'Error find person to add gift.', data: err });
+        var gift = new Gift(req.body);
+        gift.save(function (err, gift) {
+            if (err) return res.status(500).json({ msg: 'Error adding gift.', data: err });
+            person.gifts.push(gift._id);
+            person.save();
+            return res.status(200).json({ data: gift });
+        });
     });
-
-    //Add gift id to the person
-    var id = req.params.id;
-    //
-    //
-    //
 });
 
 //Update a gift
@@ -29,17 +29,20 @@ router.patch('/people/:id/gifts/:giftId', function (req, res) {
 
 //Delete a gift
 router.delete('/people/:id/gifts/:giftId', function (req, res) {
-    Gift.findOneAndDelete({ _id: req.params.giftId },
-        function (err, gift) {
-            if (err) return res.status(500).json({ msg: 'Error deleting gift.', data: err });
-            return res.status(200).json({ msg: 'Deleted', data: gift });
+    //Remove gift and the id from the person
+    People.findById(req.params.id, (err, person) => {
+        if (err || person==null) return res.status(500).json({ msg: 'Error find person to add gift.', data: err });
+        Gift.findOneAndDelete({ _id: req.params.giftId },
+            function (err, gift) {
+                if (err) return res.status(500).json({ msg: 'Error deleting gift.', data: err });
+                const giftIDX = person.gifts.indexOf(gift._id);
+                if (giftIDX > -1) {
+                    person.gifts.splice(giftIDX, 1);
+                    person.save();
+                }
+                return res.status(200).json({ msg: 'Deleted', data: gift });
         });
-
-    //Remove gift id from the person
-    var id = req.params.id;
-    //
-    //
-    //
+    });
 });
 
 module.exports = router;
